feat: 添加 tooltip 配置以支持矩形显示类型，并修复价格比较逻辑中的类型转换

+++ src/view/CandleLastPriceLabelView.ts
@@ -40,7 +40,6 @@ export default class CandleLastPriceLabelView extends View {
         const { close, open } = data
         const comparePrice = lastPriceMarkStyles.compareRule === CandleColorCompareRule.CurrentOpen ? open : (dataList[dataList.length - 2]?.close ?? close)
         const priceY = yAxis.convertToNicePixel(close)
-        // console.log(pane.getYAxisWidget())
         let backgroundColor = ''
         if (close > comparePrice) {
           backgroundColor = lastPriceMarkStyles.upColor
@@ -69,7 +68,7 @@ export default class CandleLastPriceLabelView extends View {
         }
 
         if (yAxis.name === 'percentage') {
-          backgroundColor = Number.parseInt(text) > 0 ? lastPriceMarkStyles.upColor : lastPriceMarkStyles.downColor
+          backgroundColor = Number.parseFloat(text) > 0 ? lastPriceMarkStyles.upColor : lastPriceMarkStyles.downColor
         }
 
         this.createFigure({
